feat #258: create paws component
