+++ components/atoms/paws/styled.ts
@@ -1,8 +1,8 @@
 import styled from 'styled-components'
 
-export const Col = styled.div`
+export const Row = styled.div`
   align-items: center;
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
   gap: 8px;
 `
